rename to pynvml_utils
